fix(ci): keep generator intake client-safe
